requirements.txt
tensorflow
fastapi
uvicorn
streamlit
Pillow
opencv-python
numpy
skimage
requests